♻️ refactor: 미분류 아이콘 컴포넌트 하위 폴더로 분류
최상위에 방치되던 아이콘 컴포넌트를 의미에 맞는 하위 폴더로 분류한다. RubberConeIcon은 동일한 라바콘인 error-cone.svg가 있는 icon/error로 이동하고 Toast import 경로를 갱신한다. Favicon은 UMC 마크 계열인 icon/logo로 이동한다.
