+++ src/shared/assets/icon/error/RubberConeIcon.tsx
